requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.18
google-api-python-client==2.110.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==2.0.0
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
sentence-transformers==2.2.2